requirements.txt
streamlit
easyocr
groq
pillow
numpy